+++ keyboards/helix/rev1/keymaps/OLED_sample/serial_config.h
@@ -8,6 +8,9 @@
 #define SERIAL_PIN_MASK _BV(PD2)
 #define SERIAL_PIN_INTERRUPT INT2_vect
 
+#define SERIAL_SLAVE_BUFFER_LENGTH MATRIX_ROWS/2
+#define SERIAL_MASTER_BUFFER_LENGTH 0
+
 //// #error rev1/keymaps/OLED_sample serial config
 
 #endif /* SOFT_SERIAL_CONFIG_H */
