+++ keyboards/helix/rev1/rules.mk
@@ -1,5 +1,4 @@
 SRC += rev1/matrix.c
 SRC += rev1/split_util.c
-SRC += rev1/split_scomm.c
 
 BACKLIGHT_ENABLE = no
