+++ keyboards/helix/rev2/keymaps/five_rows_jis/config.h
@@ -24,7 +24,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // place overrides here
 
 // if you need more program area, try uncomment follow line
-//#undef SERIAL_USE_MULTI_TRANSACTION
+//#include "serial_config_simpleapi.h"
 
 #ifdef MOUSEKEY_ENABLE
   #undef MOUSEKEY_INTERVAL
