+++ keyboards/helix/rev2/keymaps/led_test/config.h
@@ -22,7 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define CONFIG_USER_H
 
 // if you need more program area, try uncomment follow line
-#undef SERIAL_USE_MULTI_TRANSACTION
+#include "serial_config_simpleapi.h"
 
 // place overrides here
